note: guard flock in NextID behind unix build tag
